change num type to u16

+++ src/main.rs
@@ -22,10 +22,10 @@ fn main() {
 
   let graph = read_graph(DATA);
   let word_to_id_map: WordToIDMap = make_word_to_id_map(&graph);
-  let id: i32 = *word_to_id_map.get(word).unwrap();
+  let id: u16 = *word_to_id_map.get(word).unwrap();
   let id_to_next_map = make_id_to_node_map(&graph);
   let now = timestamp();
-  let res: Vec<i32> = find_longest_chain(
+  let res: Vec<u16> = find_longest_chain(
     id,
     vec![id],
     &id_to_next_map,
